+++ src/keybinds/config.rs
@@ -132,10 +132,13 @@ pub struct LogTabKeybindsConfig {
     pub copy_rev: Option<Keybind>,
     pub rebase: Option<Keybind>,
 
+    pub push_menu: Option<Keybind>,
     pub push: Option<Keybind>,
     pub push_new: Option<Keybind>,
     pub push_all: Option<Keybind>,
     pub push_all_new: Option<Keybind>,
+    pub push_change: Option<Keybind>,
+    pub push_named: Option<Keybind>,
     pub fetch: Option<Keybind>,
     pub fetch_all: Option<Keybind>,
 }